requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==4.2.19
django-anymail==14.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
mysqlclient==2.2.8
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
django-environ==0.13.0
whitenoise==6.12.0
waitress==3.0.2
